fixing type cast error in xen/common/sched_rt.c

+++ xen/common/sched_rt.c
@@ -1247,7 +1247,7 @@ rt_dom_cntl(
                           budget < MICROSECS(10) || budget > period )
             {
                 printk("RTDS: period (%ld) or budget (%ld) is invalid",
-                        period, budget);
+                        (long int) period, (long int) budget);
                 rc = -EINVAL;
                 break;
             }
